Propagate the py.typed marker to the wheel (#21)
The [previous pull
request](https://github.com/OpenBB-finance/openbb-ai/pull/15) that
introduced the `py.typed` marker to satisfy mypy did not correctly
propagate it into the package wheel. This results in mypy still
complaining if the package is installed from pypi.

This PR adds poetry config to include the marker into the wheel.